Remove all index mentions from error messages in reports
If someone wants to process rows or columns indexes, e.g. choose
between zero and one based indexes, including the index in the
message complicates things. The index is already included in a
separate value in the report so it can always be included and
processed as needed.

This still leaves the indexes as zero based because we all know
those are way better.