Restructure the README documentation
I've improved the readability of our README, focusing on using goodtables as a
CLI tool, not as a Python module. I moved the `validate()` docstrings to the
function itself. It's quite messy, as the arguments depend on the preset of the
data. As far as I could see, there's no standard way of defining our arguments
in docstrings. In the future, we might end up splitting it into multiple
methods, `validate_table()`, `validate_datapackage()`, etc, each with its
specific arguments.