[#233] Improve CLI docs and allow disabling row/table/error limits
The user is able to disable the limits by setting it to -1. The main change
are:

* Remove `infer_fields` option
    We only have `infer_schema` now. The behaviour of `infer_fields` can be
used by defining an explicit schema and enabling `infer_schema`, which will
only infer the fields not present in the explicit schema. One less parameter to
document, and for users to understand.

* Renamed `order_fields` to `ignore_order`
    The focus is on the user intent. Order fields talks about how we
implemented it internally, but what the user cares about is that her data is
valid even if the column order is different.

* Added `--quiet` option
    We don't output anything if this option is set.