Extract error dict into Error class and don't modify received "errors" argument
By using a class, we keep the knowledge on what fields the error dict
contains into a single place, making it easier both to document the format, and
to modify it if needed.

Regarding the "errors" argument, I modified the code to return the errors found
instead of appending to an output variable. This makes the functions closer to
being idempotent (some still modify the "cells" array), and are more Pythonic
overall.