Move cell creation to goodtables.cells and simplify @check arguments
This reduces the repetition on cell creation, and creates a single place to add
functions that handle cells (e.g. `is_complete()`).

Initially, I created a `Cell` class to handle this. However, as noted by @roll,
this creates performance issues as we're creating an object for each cell in a
potentially large table. Using built-in structures is faster.

I also added the `row-number` to the cell itself, and renamed `number` to
`column-number`. To avoid breaking other people's codes, `number` still exists,
but our code only uses `column-number`. We can remove it in the future.

Regarding @check, I modified it to not need to pass the `type` and `context` on
the default checks, as they're already defined in the spec.json. This should
have been a separate commit.