Simplify error, cells, and checks signature, and add report validation JSON Schema
* Extract error dict into Error class and don't modify received "errors" argument

By using a class, we keep the knowledge on what fields the error dict
contains into a single place, making it easier both to document the format, and
to modify it if needed.

Regarding the "errors" argument, I modified the code to return the errors found
instead of appending to an output variable. This makes the functions closer to
being idempotent (some still modify the "cells" array), and are more Pythonic
overall.

* Add JSON Schema for validation report

It allow us not only to validate the report, but can also be used as a basis to
document it.

* Move cell creation to goodtables.cells and simplify @check arguments

This reduces the repetition on cell creation, and creates a single place to add
functions that handle cells.

Initially, I created a `Cell` class to handle this. However, as noted by @roll,
this creates performance issues as we're creating an object for each cell in a
potentially large table. Using built-in structures is faster.

I also added the `row-number` to the cell itself, and renamed `number` to
`column-number`. To avoid breaking other people's codes, `number` still exists,
but our code only uses `column-number`. We can remove it in the future.

Regarding @check, I modified it to not need to pass the `type` and `context` on
the default checks, as they're already defined in the spec.json. This should
have been a separate commit.