[#233] Improve CLI and add init option to generate datapackage
* Improve CLI docs and allow disabling row/table/error limits

The user is able to disable the limits by setting it to -1. The main change
are:

* Remove `infer_fields` option
    We only have `infer_schema` now. The behaviour of `infer_fields` can be
used by defining an explicit schema and enabling `infer_schema`, which will
only infer the fields not present in the explicit schema. One less parameter to
document, and for users to understand.

* Added `--quiet` option
    We don't output anything if this option is set.

* Accept multiple sources via CLI

* Add CLI option to initialize datapackage with schemas

The output of `goodtables init my-data.csv` is a Tabular Data Package
descriptor with the table schema inferred from the data. This makes it easier
to start using goodtables, because the user will just need to tweak the
descriptor.

* Add --output option to redirect output to file