[#251] Allow non-required primary keys
If a field is a `primaryKey`, we validate that it exists, the same way as if it
had `{'constraints': {'required': True}}`. However, there can be some cases
where a `primaryKey` field can be null. Pretty uncommon, but messy data...

This commit allows removing the check by explicitly setting the `required` to
False, so a field like:

```
"fields": [
  {
    "name": "id",
    "constraints": {
      "required": False
    }
  }
],
"primaryKey": "id"
```

Wouldn't be required. In other words, the `required` constraint overloads the
`primaryKey` constraint.

Fixes #251