Fix typos, grammar in docs.
This commit adds fixes to the docs for typos and grammatical mistakes.
Most changes affect only one line, but many also span across sentences
and may have the potential to change the meaning of the documentation.
Changes to the meaning/interpretation of the docs is not intentional.

This also fixes the label of a duplicated kwarg in `describe_resource`.

encoding --> hashing