1631: Pin black, moto, pytest versions to avoid breaking changes to tests
# Summary

Over the weekend (26-27 Jan 2024) there were major updates to black, moto, and pytest that broke the tests.  This commit pins the versions of those tools to before their major updates to get the automated running & passing again.

Actual updates to the code to be in line with the latest versions can follow in a later PR.

# Tests

Run `hatch -e ci.py3.10 run test` and ensure all tests run and pass (with test db images up and running).